requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.1.1
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
